+++ src/pygama/flow/__init__.py
@@ -4,6 +4,6 @@ Subpackage description
 
 from .data_loader import DataLoader
 from .datagroup import DataGroup
-from .fileDB import FileDB
+from .file_db import FileDB
 
 __all__ = ["DataGroup", "DataLoader", "FileDB"]
